pyproject.toml
[project]
name = "project1-energy-analysis"
version = "0.1.0"
description = "A Streamlit project combining NOAA weather data and EIA electricity data."
authors = [
    { name = "Developer" }
]
requires-python = ">=3.8"
dependencies = [
    "streamlit>=1.20.0",
    "pandas>=2.0.0",
    "requests>=2.28.0",
    "plotly>=5.10.0",
    "python-dotenv>=1.0.0",
    "scipy>=1.10.0",
    "pyyaml>=6.0"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
